+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to INTERSECT will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
+### Changed
+- Action tooltips are shorter, and RESEQ now has a Shift+R shortcut.
 
 ## [0.10.0] - 2026-03-20
 ### Added